shithead; a tool to do bulk whois lookups over tor.
you shouldn't do this. it violates terms-of-service for WHOIS providers. It's also probably not terribly friendly to the Tor
network (although, it does inject a bunch of extra junk traffic, so who knows, maybe it frustrates people monitoring tor just a little?)
But really. you shouldn't do this.

socks.py is by Dan Haim. it's BSD licensed, and you can find it at socksipy.sourceforge.net if you want.